requirements.txt
deepseek-ai>=0.1.0
neo4j>=5.14.0
networkx>=3.2.1
pyyaml>=6.0.1
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
selenium>=4.16.0
pytube>=15.0.0
transformers>=4.36.0
torch>=2.1.0
gitpython>=3.1.40
pydantic>=2.5.2
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
aiohttp>=3.9.1
pytest>=7.4.3
black>=23.12.1
isort>=5.13.2
mypy>=1.8.0 